% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartowp-package.R
\docType{package}
\name{quartowp-package}
\alias{quartowp}
\alias{quartowp-package}
\title{quartowp: Publish Quarto Articles to WordPress}
\description{
The quartowp package renders Quarto documents and publishes their text,
tables, and figures to WordPress through the WordPress REST API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/petzi53/quartowp}
  \item Report bugs at \url{https://github.com/petzi53/quartowp/issues}
}

}
\author{
\strong{Maintainer}: Peter Baumgartner \email{petzi53@gmail.com}

Authors:
\itemize{
  \item Peter Baumgartner \email{petzi53@gmail.com}
}

}
\keyword{internal}
