% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_publish.R
\name{wp_publish}
\alias{wp_publish}
\title{Publish a Quarto document to WordPress}
\usage{
wp_publish(input = NULL, status = "draft")
}
\arguments{
\item{input}{Path to a Quarto document. If \code{NULL}, the currently active
source document in RStudio or Positron is used via \code{rstudioapi}.}

\item{status}{WordPress post status.}
}
\value{
An object of class \code{quartowp_post}.
}
\description{
Publish a Quarto document to WordPress
}
